- Marten added some static type-checking to SCF. - Marten fixed VC8 compatibility for rvl_s latest commit. - Marten added render buffer saving to genmesh factory saver. - Marten fixed some random warnings and errors in comments.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@23448 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csgeom/vector3.h
@@ -208,16 +208,16 @@ public:
 
   /// Returns n-th component of the vector.
 #ifdef __STRICT_ANSI__
-  inline float operator[] (int n) const { return !n?x:n&1?y:z; }
+  inline float operator[] (size_t n) const { return !n?x:n&1?y:z; }
 #else
-  inline float operator[] (int n) const { return m[n]; }
+  inline float operator[] (size_t n) const { return m[n]; }
 #endif
 
   /// Returns n-th component of the vector.
 #ifdef __STRICT_ANSI__
-  inline float & operator[] (int n) { return !n?x:n&1?y:z; }
+  inline float & operator[] (size_t n) { return !n?x:n&1?y:z; }
 #else
-  inline float & operator[] (int n) { return m[n]; }
+  inline float & operator[] (size_t n) { return m[n]; }
 #endif
 
   /// Add another vector to this vector.
